+++ Utilities/MathUtil.cs
@@ -29,7 +29,7 @@ namespace Expanse.Utilities
         /// <param name="value">Int value to perform a modulo operation on.</param>
         /// <param name="mod">Modulo value.</param>
         /// <returns>Returns the modulo result from the operation.</returns>
-        public static float Modulo(int value, int mod)
+        public static int Modulo(int value, int mod)
         {
             return (value % mod + mod) % mod;
         }
